:sparkles: remove super custom code
In previous versions of Doctrine, it was hard to act on entity load.
Since doctrine changed its internal behavior, we managed to improve the
situation as well on the bundle.

The new PostLoadDispatcherInjecttionListener class allows us to remove a
huge part of this bundle that was previously super-mega-custom.

This is awesome! But it comes with a huge refactoring, and this is not
even the last commit toward the next major version of this package.